issue #336: lseek / pread64 / pwrite64 / copy_file_range の off_t を long 化 (2GB 超ファイル)
FileAccess.FileSeek が offset / 内部位置 (o/size/curptr) / 戻り値を int で持ち、
RandomAccessFile の getFilePointer()/length() を (int) に丸めていたため、off_t
(64-bit) が 2GB (2^31) 境界で切り詰められていた。lseek だけでなく FileSeek を
経由する pread64 / pwrite64 / copy_file_range も同じ欠陥を共有 (各所 (int)offset)。

修正:
- FileSeek を long FileSeek(int fd, long offset, int whence) に。内部 o/size/curptr
  を long、getFilePointer()/length() の (int) cast を除去、seek(long) をそのまま使用。
  dir 分岐の set_ptr(int) には (int)o (dir cookie は小なので可)。
- sys_lseek (Syscall) / amd64_pread64 / amd64_pwrite64 / amd64_copy_file_range の
  (int)offset / (int)in_off / (int)out_off cast を除去、saved/in_saved/out_saved を long に。

回帰: 新規 sys_lseek_large64 (4GiB+5 で lseek SET/END + write/read roundtrip +
pwrite64/pread64 + hole 確認、sparse file) PASS。sys_lseek64 / sys_pwrite64 /
sys_copy_file_range64 / hello_dyn64 (ld.so の libc file-backed mmap = 内部 FileSeek) /
fileio64 / mmap_dyn64 PASS。

残 (#336 内): mmap の file offset 経路 (alloc_and_map の int offset / AllocInfo.map_offset)
は MemoryBackend interface 波及のため別 commit/PR とする。

+++ src/main/java/emulin/Syscall.java
@@ -792,7 +792,7 @@ public class Syscall extends EmuSocket
     int fd = (int)bx;
     long offset = cx;  // off_t は 64-bit
     int whence = (int)dx;
-    return( FileSeek( fd, (int)offset, whence ));
+    return( FileSeek( fd, offset, whence ));   // issue #336: off_t を切り詰めない
   }
   long sys_getpid( long bx, long cx, long dx, long si, long di ) {    return( process.pid );  }
   long sys_mount( long bx, long cx, long dx, long si, long di )  {
